+++ src/pyplier/utils.py
@@ -1,4 +1,3 @@
-from functools import singledispatch
 from typing import Any, Optional, Union
 
 import numpy as np
@@ -31,34 +30,11 @@ def tcrossprod(
         return mat1 @ mat2.transpose()
 
 
-@singledispatch
-def rowNorm(arr):
-    return arr
-
-
-@rowNorm.register
-def _(arr: np.ndarray) -> np.ndarray:
-    return np.apply_along_axis(lambda x: ((x - np.mean(x)) / np.std(x, ddof=1)), 1, arr)
-
-
-@rowNorm.register
-def _(arr: pd.DataFrame) -> pd.DataFrame:
-    return arr.apply(lambda x: ((x - np.mean(x)) / np.std(x, ddof=1)), axis=1)
-
-
-@singledispatch
-def colNorm(arr):
-    return arr
-
-
-@colNorm.register
-def _(arr: np.ndarray) -> np.ndarray:
-    return np.apply_along_axis(lambda x: ((x - np.mean(x)) / np.std(x, ddof=1)), 0, arr)
-
-
-@colNorm.register
-def _(arr: pd.DataFrame) -> pd.DataFrame:
-    return arr.apply(lambda x: ((x - np.mean(x)) / np.std(x, ddof=1)), axis=0)
+def zscore(arr: Union[pd.Series, np.ndarray]) -> Union[pd.Series, np.ndarray]:
+    if np.std(arr, ddof=1) == 0.0:
+        return np.zeros(len(arr))
+    else:
+        return (arr - np.mean(arr)) / np.std(arr, ddof=1)
 
 
 def setdiff(list1: list[Any], list2: list[Any]) -> list[Any]: